Separate wlr_screencopy code into separate file
