+++ include/wlr_screencast.h
@@ -5,9 +5,6 @@
 
 #define WL_OUTPUT_VERSION 4
 
-#define SC_MANAGER_VERSION 3
-#define SC_MANAGER_VERSION_MIN 2
-
 #define WL_SHM_VERSION 1
 
 #define LINUX_DMABUF_VERSION 4
